+++ superduper/backends/base/backends.py
@@ -1,9 +1,97 @@
 import typing as t
 from abc import ABC, abstractmethod
+from collections import defaultdict
 
 if t.TYPE_CHECKING:
     from superduper.base.datalayer import Datalayer
     from superduper.components.component import Component
+
+
+class Bookkeeping(ABC):
+    """Mixin class for tracking components and associated tools."""
+
+    def __init__(self):
+        self.component_uuid_mapping = defaultdict(set)
+        self.uuid_component_mapping = {}
+        self.tool_uuid_mapping = defaultdict(set)
+        self.uuid_tool_mapping = {}
+        self.tools = {}
+
+    def build_tool(self, component: 'Component'):
+        """Build a tool from a component.
+
+        :param component: Component to build tool from.
+        """
+        pass
+
+    def get_tool(self, uuid: str):
+        """Get the tool from a uuid.
+
+        :param uuid: UUID of the tool.
+        """
+        tool_id = self.uuid_tool_mapping[uuid]
+        return self.tools[tool_id]
+
+    def put_component(self, component: 'Component', **kwargs):
+        """Put a component to the backend.
+
+        :param component: Component to put.
+        :param kwargs: kwargs dictionary.
+        """
+        tool = self.build_tool(component)
+        tool.db = self.db
+        self.component_uuid_mapping[(component.component, component.identifier)].add(
+            component.uuid
+        )
+        self.uuid_component_mapping[component.uuid] = (
+            component.component,
+            component.identifier,
+        )
+        self.uuid_tool_mapping[component.uuid] = tool.identifier
+        self.tool_uuid_mapping[tool.identifier].add(component.uuid)
+        self.tools[tool.identifier] = tool
+        tool.initialize(**kwargs)
+
+    def drop_component(self, component: str, identifier: str):
+        """Drop the component from backend.
+
+        :param component: Component name.
+        :param identifier: Component identifier.
+        """
+        uuids = self.component_uuid_mapping[(component, identifier)]
+        tool_ids = []
+        for uuid in uuids:
+            del self.uuid_component_mapping[uuid]
+            tool_id = self.uuid_tool_mapping[uuid]
+            tool_ids.append(tool_id)
+            del self.uuid_tool_mapping[uuid]
+            self.tool_uuid_mapping[tool_id].remove(uuid)
+            if not self.tool_uuid_mapping[tool_id]:
+                self.tools[tool_id].drop()
+                del self.tools[tool_id]
+        del self.component_uuid_mapping[(component, identifier)]
+
+    def drop(self):
+        """Drop the backend."""
+        for tool in self.tools.values():
+            tool.drop()
+        self.component_uuid_mapping = defaultdict(set)
+        self.uuid_component_mapping = {}
+        self.tool_uuid_mapping = defaultdict(set)
+        self.uuid_tool_mapping = {}
+        self.tools = {}
+
+    def list_components(self):
+        """List components, and identifiers deployed."""
+        return list(self.component_uuid_mapping.keys())
+
+    def list_tools(self):
+        """List tools deployed."""
+        return list(self.tools.keys())
+
+    def list_uuids(self):
+        """List uuids deployed."""
+        return list(self.uuid_component_mapping.keys())
 
 
 class BaseBackend(ABC):
@@ -34,34 +122,19 @@ class BaseBackend(ABC):
         """To be called on program start."""
         pass
 
-    def put_component(self, component: 'Component', **kwargs):
+    @abstractmethod
+    def put_component(self, component: 'Component'):
         """Add a component to the deployment.
 
         :param component: ``Component`` to put.
-        :param kwargs: kwargs dictionary.
         """
-        # This is to make sure that we only have 1 version
-        # of each component implemented at any given time
-        # TODO: get identifier in string component argument.
-        identifier = ''
-        if isinstance(component, str):
-            uuid = component
-        else:
-            uuid = component.uuid
-            identifier = component.identifier
-
-        if uuid in self.list_uuids():
-            return
-        if identifier in self.list_components():
-            del self[component.identifier]
-
-        self._put(component, **kwargs)
 
     @abstractmethod
     def drop_component(self, component: str, identifier: str):
         """Drop the component from backend.
 
-        :param identifier: Component identifier
+        :param component: Component name.
+        :param identifier: Component identifier.
         """
 
     @property