+++ superduper/backends/base/crontab.py
@@ -3,4 +3,3 @@ from superduper.backends.base.backends import BaseBackend
 
 class CrontabBackend(BaseBackend):
     """Base class for crontab backends."""
-