+++ superduper/components/component.py
@@ -1,4 +1,5 @@
 """The component module provides the base class for all components in superduper.io."""
+
 import dataclasses as dc
 import json
 import os
@@ -56,8 +57,10 @@ def _build_info_from_path(path: str):
         files = {}
         for file_id in os.listdir(os.path.join(path, "files")):
             sub_paths = os.listdir(os.path.join(path, "files", file_id))
-            assert len(sub_paths) == 1, f"Multiple files found in {file_id}"
-            file_name = sub_paths[0]
+            # assert len(sub_paths) == 1, f"Multiple files found in {file_id}"
+            file_name = next(
+                x for x in sub_paths if not x.startswith(".") or x.startswith("_")
+            )
             files[file_id] = os.path.join(path, "files", file_id, file_name)
         config_object[KEY_FILES] = files
 
@@ -464,12 +467,9 @@ class Component(Base, metaclass=ComponentMeta):
         if not self.identifier:
             raise ValueError('identifier cannot be empty or None')
 
-    def cleanup(self, db: 'Datalayer'):
-        """Method to clean the component.
-
-        :param db: The `Datalayer` to use for the operation.
-        """
-        db.cluster.cache.drop(self)
+    def cleanup(self):
+        """Method to clean the component."""
+        pass
 
     def _get_metadata(self):
         """Get metadata of the component."""
@@ -484,19 +484,16 @@ class Component(Base, metaclass=ComponentMeta):
         """Get dependencies on the component."""
         return ()
 
-    # TODO why both methods?
     def init(self):
         """Method to help initiate component field dependencies."""
-        self.unpack()
 
-    def unpack(self):
-        """Method to unpack the component.
-
-        This method is used to initialize all the fields of the component and leaf
-        """
+        def mro(item):
+            objects = item.__class__.__mro__
+            return [f'{o.__module__}.{o.__name__}' for o in objects]
 
         def _init(item):
-            if isinstance(item, Component):
+
+            if 'superduper.components.component.Component' in mro(item):
                 item.init()
                 return item
 
@@ -534,30 +531,24 @@ class Component(Base, metaclass=ComponentMeta):
             child.pre_create(db)
         self._pre_create(db)
 
-    def on_create(self, db: 'Datalayer') -> None:
+    def on_create(self):
         """Called after the first time this component is created.
 
         Generally used if ``self.version`` is important in this logic.
-
-        :param db: the db that creates the component.
         """
-        assert db
-        self.declare_component(db.cluster)
+        assert self.db is not None
+        self.declare_component()
 
-    def declare_component(self, cluster):
-        """Declare the component to the cluster.
-
-        :param cluster: The cluster to declare the component to.
-        """
+    def declare_component(self):
+        """Declare the component to the cluster."""
         pass
 
     @staticmethod
-    def read(path: str, db: t.Optional['Datalayer'] = None):
+    def read(path: str):
         """
         Read a `Component` instance from a directory created with `.export`.
 
         :param path: Path to the directory containing the component.
-        :param db: Datalayer instance to be used to read the component.
 
         Expected directory structure:
         ```
@@ -566,45 +557,8 @@ class Component(Base, metaclass=ComponentMeta):
         |_files/*
         ```
         """
-        was_zipped = False
-        if path.endswith('.zip'):
-            was_zipped = True
-            import shutil
-
-            shutil.unpack_archive(path)
-            path = path.replace('.zip', '')
-
         config_object = _build_info_from_path(path=path)
-
-        from superduper import Document
-
-        if db is not None:
-            for blob in os.listdir(path + '/' + 'blobs'):
-                with open(path + '/blobs/' + blob, 'rb') as f:
-                    data = f.read()
-                    db.artifact_store.put_bytes(data, blob)
-
-            out = Document.decode(config_object, db=db).unpack()
-        else:
-            from superduper.base.artifacts import FileSystemArtifactStore
-
-            artifact_store = FileSystemArtifactStore(
-                conn=path,
-                name='tmp_artifact_store',
-                files='files',
-                blobs='blobs',
-            )
-            db = namedtuple('tmp_db', field_names=('artifact_store',))(
-                artifact_store=artifact_store
-            )
-            cls = import_object(config_object['_path'])
-            out = Document.decode(
-                config_object, schema=cls.class_schema, db=db
-            ).unpack()
-            out = cls.from_dict(out, db=db)
-            if was_zipped:
-                shutil.rmtree(path)
-        return out
+        return Component.decode(config_object)
 
     def export(
         self,