+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use databackend to perform metadata duties
 - Add `db.create` and `db.insert` instead of `auto_schema`
 - Merkel-tree implementation replacing random `.uuid` with deterministic implementation
+- Simplify the `Template` class
 
 #### New Features & Functionality